+++ tests/cmake/AddTestMaps.cmake
@@ -189,6 +189,7 @@ list(APPEND LUA_TEST_MAPS
   "bugs/1533_savegame_write_escape"
   "bugs/1534_unicode_filenames"
   "bugs/1535_collision_moving_sprite"
+  "bugs/1546_custom_entities_collisions"
   "bugs/1547_teletransporter_on_activated_twice/map_1"
   "bugs/1550_use_bow"
   "bugs/1551_error_teletransporter_set_destination_map"