+++ changelog.md
@@ -8,6 +8,7 @@
 * Fix a001 error while trying to set volume on a non-playing music (#1935).
 * Fix duplicate teletransporter activation when hero has no linked camera (#1945).
 * Fix left and right inverted when using analog controls and keyboard default mappings (#1588).
+* Fix `set_can_traverse` between two custom entities (#1546).
 
 ## Solarus 2.0.2 (2025-12-05)